Remove apt-over-https dependencies
The only dependency in this var that is still required after apt
version 1.5 is gnupg2, and even Debian Buster comes with version 1.8.

gnupg2 is not strictly an apt-over-https dependency; it is required
when adding new package repos even if they use HTTP.  So it makes
sense to install it directly in the playbook.